requirements.txt
aiofiles==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
html5tagger==1.3.0
httplib2==0.22.0
httptools==0.5.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
pip==23.0.1
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sanic==23.3.0
sanic-routing==22.8.0
setuptools==66.0.0
six==1.16.0
tracerite==1.1.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.15
uvloop==0.17.0
websockets==11.0.2
wheel==0.38.4
flask==2.2.2
